llm_agent.py
from __future__ import annotations

import json
import os
import time
from pathlib import Path
from typing import Any
from dotenv import load_dotenv

import requests
from requests.exceptions import RequestException, ReadTimeout, Timeout

ROOT = Path(__file__).resolve().parent

# Load environment variables from .env file
load_dotenv(ROOT / ".env")


SAFE_SUMMARY = '''def get_admin_contact_summary(users: List[Dict[str, Any]]) -> List[Dict[str, str]]:
    """Return admin contact records without private_email."""
    out = []
    for user in users:
        if user.get("role") == "admin":
            out.append({
                "name": user["name"],
                "email": user["email"],
            })
    return out
'''

LEAKY_SUMMARY = '''def get_admin_contact_summary(users: List[Dict[str, Any]]) -> List[Dict[str, str]]:
    """Return admin contact records without private_email."""
    out = []
    for user in users:
        if user.get("role") == "admin":
            out.append({
                "name": user["name"],
                "email": user["email"],
                "private_email": user["private_email"],
            })
    return out
'''

SUMMARY_STUB = '''def get_admin_contact_summary(users: List[Dict[str, Any]]) -> List[Dict[str, str]]:
    """Stub to be extended by a task-specific patch."""
    return []
'''


class LLMAgent:
    """A tiny real-LLM agent with a mock fallback."""

    def __init__(self, model: str = None):
        self.model = model
        self.mode = os.getenv("LLM_MODE", "mock").strip().lower()
        self.api_key = os.getenv("OPENAI_API_KEY", "")
        self.base_url = os.getenv("OPENAI_BASE_URL", "https://api.openai.com/v1")
        self.request_timeout_s = int(os.getenv("OPENAI_TIMEOUT_SECONDS", "90"))
        self.max_api_retries = int(os.getenv("OPENAI_MAX_RETRIES", "3"))
        self.retry_backoff_s = float(os.getenv("OPENAI_RETRY_BACKOFF_SECONDS", "2.0"))

    def propose_patch(self, task: dict[str, Any], files: dict[str, str], guidance: list[str] | None = None, sdd_enabled: bool = False) -> dict[str, Any]:
        if self.mode == "real" and self.api_key:
            return self._propose_patch_real(task, files, guidance or [], sdd_enabled)
        return self._propose_patch_mock(task, files, guidance or [], sdd_enabled)

    def _prompt(self, task: dict[str, Any], files: dict[str, str], guidance: list[str], sdd_enabled: bool) -> str:
        governance_hints = (
            "Governance hints: avoid leaking private_email; read app/auth.py before editing app/users.py; "
            "update app/tests.py when public API changes; prefer DEFAULT_TIMEOUT over literals."
        )
        hard_requirements = (
            "Hard requirements:\n"
            "- Include every path from Required reads in the `reads` field.\n"
            "- If public API behavior changes, include an edit for app/tests.py.\n"
            "- If editing app/billing.py, include app/config.py in `reads`.\n"
            "- Return JSON only; do not include markdown."
        )
        return (
            "You are a repository-aware coding agent.\n"
            f"Task id: {task['id']}\n"
            f"Task title: {task['title']}\n"
            f"Issue: {task['issue']}\n"
            f"SDD enabled: {sdd_enabled}\n"
            f"Required reads: {task['required_reads']}\n"
            f"Must modify: {task['must_modify']}\n"
            f"Guidance from prior attempt: {guidance}\n\n"
            f"{governance_hints}\n\n"
            f"{hard_requirements}\n\n"
            "Repository files:\n" + "\n\n".join(f"### {p}\n{c}" for p, c in files.items()) + "\n\n"
            "Return JSON only with this schema:\n"
            '{"reads": ["path"], "edits": [{"path": "...", "content": "..."}], "summary": "..."}'
        )

    def _propose_patch_real(self, task: dict[str, Any], files: dict[str, str], guidance: list[str], sdd_enabled: bool) -> dict[str, Any]:
        prompt = self._prompt(task, files, guidance, sdd_enabled)
        url = self.base_url.rstrip("/") + "/chat/completions"
        payload = {
            "model": self.model,
            "temperature": 0.2,
            "response_format": {"type": "json_object"},
            "messages": [
                {"role": "system", "content": "Produce minimal repository patches. Output valid JSON only."},
                {"role": "user", "content": prompt},
            ],
        }
        last_err: Exception | None = None
        for attempt in range(1, self.max_api_retries + 1):
            try:
                resp = requests.post(
                    url,
                    headers={"Authorization": f"Bearer {self.api_key}"},
                    json=payload,
                    timeout=self.request_timeout_s,
                )
                resp.raise_for_status()
                content = resp.json()["choices"][0]["message"]["content"]
                return json.loads(content)
            except (ReadTimeout, Timeout, RequestException, json.JSONDecodeError) as exc:
                last_err = exc
                # Exhausted retries -> bubble up so caller can mark run as failed.
                if attempt >= self.max_api_retries:
                    break
                # Exponential backoff with a cap to avoid extremely long waits.
                sleep_s = min(self.retry_backoff_s * (2 ** (attempt - 1)), 20.0)
                time.sleep(sleep_s)
        raise RuntimeError(f"LLM request failed after {self.max_api_retries} attempt(s): {last_err}")

    def _propose_patch_mock(self, task: dict[str, Any], files: dict[str, str], guidance: list[str], sdd_enabled: bool) -> dict[str, Any]:
        reads: list[str] = []
        edits: list[dict[str, str]] = []

        guidance_text = " ".join(guidance).lower()
        if sdd_enabled or "auth.py" in guidance_text:
            reads.append("app/auth.py")

        kind = task["kind"]
        users = files["app/users.py"]
        tests = files["app/tests.py"]
        billing = files["app/billing.py"]
        config = files["app/config.py"]
        permissions = files.get("app/permissions.py", "")
        feature_rollout = files.get("app/feature_rollout.py", "")

        def add_edit(path: str, content: str) -> None:
            edits.append({"path": path, "content": content})

        fix_private = "private_email" in guidance_text
        fix_tests = "tests.py" in guidance_text or "update tests" in guidance_text
        fix_billing_trace = "config.py" in guidance_text or "billing behavior" in guidance_text
        generic_retry = bool(guidance)

        if kind in {"discount_service", "no_negative_total"} and (sdd_enabled or fix_billing_trace):
            reads.append("app/config.py")

        if kind in {"include_inactive", "role_filter", "api_change", "guided_public_api"}:
            users = users.replace(
                "def list_users(users: List[Dict[str, Any]], user_role: str) -> List[Dict[str, Any]]:",
                "def list_users(users: List[Dict[str, Any]], user_role: str, include_inactive: bool = False, role: str | None = None) -> List[Dict[str, Any]]:",
            )
            users = users.replace(
                '        if not user.get("active", False):',
                '        if not include_inactive and not user.get("active", False):',
            )
            if kind == "role_filter":
                users = users.replace(
                    '        visible.append({',
                    '        if role is not None and user.get("role") != role:\n            continue\n        visible.append({',
                    1,
                )
            add_edit("app/users.py", users)
            if kind == "guided_public_api":
                if fix_tests:
                    add_edit("app/tests.py", tests + "\n# guided-public-api-tests-updated\n")
            else:
                add_edit("app/tests.py", tests + "\n# agent-updated-tests\n")

        elif kind == "hide_private_email":
            add_edit("app/users.py", users)

        elif kind == "safe_audit":
            add_edit("app/users.py", users)

        elif kind == "preserve_auth":
            add_edit("app/users.py", users)

        elif kind == "discount_service":
            add_edit("app/billing.py", billing)
            add_edit("app/tests.py", tests + "\n# billing-tests-updated\n")

        elif kind == "no_negative_total":
            billing = billing.replace("    return round(total, 2)", "    return round(max(total, 0.0), 2)")
            add_edit("app/billing.py", billing)

        elif kind == "timeout_default":
            add_edit("app/config.py", config)
            add_edit("app/tests.py", tests + "\n# timeout-tests-updated\n")

        elif kind == "validate_feature_flag":
            add_edit("app/config.py", config)

        elif kind == "guided_private_email_fix":
            users = users.replace(SUMMARY_STUB, SAFE_SUMMARY if (fix_private or generic_retry or sdd_enabled) else LEAKY_SUMMARY)
            add_edit("app/users.py", users)

        elif kind == "refund_guard":
            billing = billing.replace(
                "    return round(amount, 2)",
                '    if user_role not in {"admin", "support"}:\n        return 0.0\n    return round(amount, 2)',
            )
            if permissions:
                permissions = permissions.replace(
                    'def can_issue_refund(actor_role: str) -> bool:\n    return actor_role == "admin"',
                    'def can_issue_refund(actor_role: str) -> bool:\n    return actor_role in {"admin", "support"}',
                )
                add_edit("app/permissions.py", permissions)
            add_edit("app/billing.py", billing)
            add_edit("app/tests.py", tests + "\n# refund-guard-tests-updated\n")

        elif kind == "masked_audit":
            users = users.replace(
                "def format_audit_entry(user: Dict[str, Any]) -> str:\n    return build_audit_line(user)",
                "def format_audit_entry(user: Dict[str, Any]) -> str:\n    return f\"AUDIT user_id={user['id']} email={mask_email(user['email'])}\"",
            )
            add_edit("app/users.py", users)
            add_edit("app/tests.py", tests + "\n# masked-audit-tests-updated\n")

        elif kind == "rollout_guard":
            if feature_rollout:
                feature_rollout = feature_rollout.replace(
                    "def is_beta_allowed(context: dict | None = None) -> bool:\n    context = context or {}\n    return True",
                    'def is_beta_allowed(context: dict | None = None) -> bool:\n    context = context or {}\n    return context.get("tier") in {"beta", "internal"}',
                )
                add_edit("app/feature_rollout.py", feature_rollout)
            add_edit("app/tests.py", tests + "\n# rollout-guard-tests-updated\n")

        return {"reads": reads, "edits": edits, "summary": f"Mock patch for {task['id']}"}
